requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
python-dotenv==1.0.1
django-autoslug==1.9.9
python-jose==3.3.0
django-cors-headers==4.4.0
drf-yasg==1.21.7
setuptools==80.0.0
